refactor: users.entity 경로 수정
역할과 책임에 따른 디렉토리 분리를 위함.
