+++ src/modules/mappings/attendees/attendees.entity.ts
@@ -1,6 +1,6 @@
 import { BaseEntity } from "src/common/entities/base.entity";
 import { Plan } from "src/modules/plans/plans.entity";
-import { User } from "src/modules/users/users.entity";
+import { User } from "src/modules/users/entities/users.entity";
 import { Entity, ManyToOne } from "typeorm";
 
 @Entity()
