+++ src/modules/mappings/userProjects/userProjects.entity.ts
@@ -1,7 +1,7 @@
 import { BaseEntity } from "src/common/entities/base.entity";
 import { projectPermission } from "src/common/enums/projectPermission.enum";
 import { Project } from "src/modules/projects/entities/projects.entity";
-import { User } from "src/modules/users/users.entity";
+import { User } from "src/modules/users/entities/users.entity";
 import { Column, Entity, ManyToOne } from "typeorm";
 
 @Entity()
