+++ src/modules/users/entities/users.entity.ts
@@ -1,15 +1,15 @@
 import { Entity, Column, OneToMany } from 'typeorm';
-import { Comment } from '../comments/comments.entity';
+import { Comment } from '../../comments/comments.entity';
 import { BaseEntity } from 'src/common/entities/base.entity';
-import { PersonalRecall } from '../personalRecalls/entities/personalRecalls.entity';
-import { UserProject } from '../mappings/userProjects/userProjects.entity';
-import { Manager } from '../mappings/managers/managers.entity';
-import { TaskFile } from '../mappings/taskFiles/taskFiles.entity';
-import { ProjectFile } from '../mappings/projectFiles/projectFiles.entity';
-import { Writer } from '../mappings/writers/writers.entity';
-import { Attendee } from '../mappings/attendees/attendees.entity';
-import { MasterPortfolio } from '../masterPortfolios/masterPortfolios.entity';
-import { FinalPortfolio } from '../finalPortfolios/finalPortfolios.entity';
+import { PersonalRecall } from '../../personalRecalls/entities/personalRecalls.entity';
+import { UserProject } from '../../mappings/userProjects/userProjects.entity';
+import { Manager } from '../../mappings/managers/managers.entity';
+import { TaskFile } from '../../mappings/taskFiles/taskFiles.entity';
+import { ProjectFile } from '../../mappings/projectFiles/projectFiles.entity';
+import { Writer } from '../../mappings/writers/writers.entity';
+import { Attendee } from '../../mappings/attendees/attendees.entity';
+import { MasterPortfolio } from '../../masterPortfolios/masterPortfolios.entity';
+import { FinalPortfolio } from '../../finalPortfolios/finalPortfolios.entity';
 
 @Entity()
 export class User extends BaseEntity {
